Still trying to resolve screen malfunction issu
figured out the problem. interrupts during screen updates, thrown of the lcd. Also, lcd.print cannot fit within the noInterrupts() timeframe,

changed keys::updateTranspose back to non-static member function, retains transposeUp and transposeDown()

updates allSoundsOff button Rgb color.

decreased potthreshold of the knob class from 5 to 4.

+++ midibuilder8/Encoder.h
@@ -5,7 +5,7 @@
 
 class Encoder {
 private:
-  volatile int8_t _delta = 0;
+  volatile int16_t _delta = 0;
   volatile long _encoderPos;
   volatile long _lastEncoded = 0;
   volatile int16_t _encoderVal = 0;
